+++ src/Dogger/Domain/Services/Provisioning/IProvisioningJob.cs
@@ -1,4 +1,5 @@
-﻿using Dogger.Domain.Services.Provisioning.Stages;
+﻿using Dogger.Domain.Services.Provisioning.Instructions.Models;
+using Dogger.Domain.Services.Provisioning.Stages;
 
 namespace Dogger.Domain.Services.Provisioning
 {
