+++ src/Dogger/Domain/Services/Provisioning/Instructions/IInstructionGroupCollector.cs
@@ -1,0 +1,12 @@
+﻿using System;
+using Dogger.Domain.Services.Provisioning.Instructions.Models;
+
+namespace Dogger.Domain.Services.Provisioning.Instructions
+{
+    public interface IInstructionGroupCollector : IDisposable
+    {
+        IInstructionGroupCollector CollectGroup(string title);
+
+        void CollectInstruction(IInstruction instruction);
+    }
+}
