+++ src/Dogger/Domain/Services/Provisioning/Instructions/Models/HttpInstructionVerb.cs
@@ -1,0 +1,11 @@
+﻿namespace Dogger.Domain.Services.Provisioning.Instructions.Models
+{
+    public enum HttpInstructionVerb
+    {
+        Get,
+        Put,
+        Post,
+        Patch,
+        Delete
+    }
+}
