+++ src/Dogger/Domain/Services/Provisioning/Instructions/Models/IInstruction.cs
@@ -1,0 +1,7 @@
+﻿namespace Dogger.Domain.Services.Provisioning.Instructions.Models
+{
+    public interface IInstruction
+    {
+        string Type { get; }
+    }
+}
