+++ src/Dogger/Domain/Services/Provisioning/Stages/IProvisioningStateContext.cs
@@ -1,4 +1,6 @@
-﻿namespace Dogger.Domain.Services.Provisioning.Stages
+﻿using Dogger.Domain.Services.Provisioning.Instructions.Models;
+
+namespace Dogger.Domain.Services.Provisioning.Stages
 {
     public interface IProvisioningStateContext
     {
